- restructuring home module
