--- TheMovieDb/Modules/Home/HomeLocalDataManager.swift
@@ -1,13 +1,0 @@
-//
-//  HomeLocalDataManager.swift
-//  TheMovieDb
-//
-//  Created by Javier Cueto on 14/11/21.
-//  
-//
-
-import Foundation
-
-class HomeLocalDataManager: HomeLocalDataManagerInputProtocol {
-    
-}
