--- TheMovieDb/Modules/Home/HomeWireFrame.swift
@@ -1,41 +1,0 @@
-//
-//  HomeWireFrame.swift
-//  TheMovieDb
-//
-//  Created by Javier Cueto on 14/11/21.
-//  
-//
-
-import Foundation
-import UIKit
-
-class HomeWireFrame: HomeWireFrameProtocol {
-
-     static func createHomeModule() -> UIViewController {
-        let view = HomeView()
-        let presenter: HomePresenterProtocol & HomeInteractorOutputProtocol = HomePresenter()
-        let interactor: HomeInteractorInputProtocol & HomeRemoteDataManagerOutputProtocol = HomeInteractor()
-        let localDataManager: HomeLocalDataManagerInputProtocol = HomeLocalDataManager()
-        let remoteDataManager: HomeRemoteDataManagerInputProtocol = HomeRemoteDataManager()
-        let wireFrame: HomeWireFrameProtocol = HomeWireFrame()
-        
-        view.presenter = presenter
-        presenter.view = view
-        presenter.wireFrame = wireFrame
-        presenter.interactor = interactor
-        interactor.presenter = presenter
-        interactor.localDatamanager = localDataManager
-        interactor.remoteDatamanager = remoteDataManager
-        remoteDataManager.remoteRequestHandler = interactor
-        remoteDataManager.service = APIService()
-        
-        return view
-    }
-    
-    func showMovie(from view: HomeViewProtocol, with movie: Movie) {
-        let controller = MovieDetailBuilder.createModule(with: movie)
-        guard let view = view as? HomeView else { return }
-        view.navigationController?.pushViewController(controller, animated: true)
-    }
-    
-}
